Remove PlanVisualizerWidget from IStrategizerEx.sln to avoid linking and compilation errors - Also fix compilation error and update Digraph some method params to be size_t

+++ Include/IStrategizer/Diagraph.h
@@ -149,7 +149,7 @@ namespace IStrategizer
             return leaves;
         }
         //---------------------------------------------------------------------------
-        int AddNode(TValue p_value, unsigned p_index)
+        size_t AddNode(TValue p_value, size_t p_index)
         {
             for (unsigned i = 0; i < _adjacencyMatrix.size(); ++i)
             {
